Move reusable game code to engine/
