+++ src/engine/containsPoint.js
