+++ src/index.html
@@ -13,7 +13,7 @@
     <body>
         <div id="canvas-container"></div>
         <button id="full-screen-button" class="reset-button">
-            <img src="./assets/images/fullscreen.svg" alt="Full screen">
+            <img src="./engine/fullscreen.svg" alt="Full screen">
         </button>
     </body>
 </html>
